+++ android/src/main/java/com/learnium/RNDeviceInfo/RNDeviceReceiver.java
@@ -8,12 +8,6 @@ import android.content.SharedPreferences;
 public class RNDeviceReceiver extends BroadcastReceiver {
   @Override
   public void onReceive(Context context, Intent intent) {
-    String action = intent.getAction();
-    if ("com.android.vending.INSTALL_REFERRER".equals(action)) {
-      SharedPreferences sharedPref = context.getSharedPreferences("react-native-device-info", Context.MODE_PRIVATE);
-      SharedPreferences.Editor editor = sharedPref.edit();
-      editor.putString("installReferrer", intent.getStringExtra("referrer"));
-      editor.apply();
-    }
+    
   }
 }